Add files via upload
The classifier is built for the iris dataset. It will work on anydataset though, provided with the name. first we have to read iris file the it generates pat file. knn_unlabeled is the main command to find the accuracy. It is a semi supervised method which have been implemented with the Knn classifier in matlab.